+++ drivers/amlogic/nand/amlnf/block/aml_nftl_block.h
@@ -54,7 +54,7 @@ struct aml_nftl_dev{
     int                   reboot_flag;
 	uint32 (*read_data)(struct aml_nftl_dev *nftl_dev, unsigned long block, unsigned nblk, unsigned char *buf);
 	uint32 (*write_data)(struct aml_nftl_dev *nftl_dev, unsigned long block, unsigned nblk, unsigned char *buf);
-    uint32 (*discard_data)(struct aml_nftl_dev *nftl_dev, unsigned long block, unsigned nblk);
+    uint32 (*discard_data)(struct aml_nftl_dev *nftl_dev, unsigned int block, uint32 nblk);
 	uint32 (*flush_write_cache)(struct aml_nftl_dev *nftl_dev);
     uint32 (*flush_discard_cache)(struct aml_nftl_dev *nftl_dev);
 
@@ -76,7 +76,7 @@ struct aml_nftl_blk{
 
 	uint32 (*read_data)(struct aml_nftl_blk *nftl_blk, unsigned long block, unsigned nblk, unsigned char *buf);
 	uint32 (*write_data)(struct aml_nftl_blk *nftl_blk, unsigned long block, unsigned nblk, unsigned char *buf);
-    uint32 (*discard_data)(struct aml_nftl_blk *nftl_blk, unsigned long block, unsigned nblk);
+    uint32 (*discard_data)(struct aml_nftl_blk *nftl_blk, unsigned int block, uint32 nblk);
 	uint32 (*flush_write_cache)(struct aml_nftl_blk *nftl_blk);
 };
